+++ src/main/java/com/example/backend/video/service/VideoService.java
@@ -2,12 +2,11 @@ package com.example.backend.video.service;
 
 import com.example.backend.customHashtag.CustomHashtag;
 import com.example.backend.customHashtag.CustomHashtagRepository;
-import com.example.backend.playlist.PlaylistService;
+import com.example.backend.playlist.service.PlaylistService;
 import com.example.backend.user.domain.User;
 import com.example.backend.video.domain.*;
 import com.example.backend.video.dto.*;
 import com.example.backend.video.repository.*;
-import com.example.backend.video.service.VideoCommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
